+++ Plugins/Smartsuit/Source/Smartsuit/Private/SmartsuitPoseNode.cpp
@@ -4,7 +4,8 @@
 #include "Smartsuit.h"
 #include "AnimationRuntime.h"
 #include "LiveLinkCustomVersion.h"
-#include "SmartsuitDefinitions.h"
+#include "LiveLinkClient.h"
+#include "RokokoSkeletonData.h"
 #include "Animation/AnimInstanceProxy.h"
 #include "Roles/LiveLinkAnimationRole.h"
 #include "SmartsuitBlueprintLibrary.h"
@@ -311,9 +312,9 @@ float FSmartsuitPoseNode::ScaleBonesToDistance(FBoneReference scaleBone, FBoneRe
 //	return FVector::ZeroVector;
 //}
 
-FQuat GetRotation2(const FName& BoneName, FSuitData* suitdata)
+FQuat GetRotation2(const FName& BoneName, FRkkActorData* suitdata)
 {
-	if (auto SmartsuitBone = suitdata->SmartsuitBones.Find(BoneName))
+	if (auto SmartsuitBone = suitdata->bones.Find(BoneName))
 	{
 		return SmartsuitBone->Uquaternion();
 	}
@@ -321,9 +322,9 @@ FQuat GetRotation2(const FName& BoneName, FSuitData* suitdata)
 	return FQuat::Identity;
 }
 
-FVector GetPosition2(const FName& BoneName, FSuitData* suitdata)
+FVector GetPosition2(const FName& BoneName, FRkkActorData* suitdata)
 {
-	if (auto SmartsuitBone = suitdata->SmartsuitBones.Find(BoneName))
+	if (auto SmartsuitBone = suitdata->bones.Find(BoneName))
 	{
 		return SmartsuitBone->UPosition();
 	}
@@ -377,7 +378,7 @@ void FSmartsuitPoseNode::EvaluateSkeletalControl_AnyThread(FComponentSpacePoseCo
 		//UE_LOG(LogTemp, Warning, TEXT("No receiver"));
 		//return;
 	}
-	FSuitData* data = nullptr;//receiver->GetSmartsuit(Controller->suitname);
+	FRkkActorData* data = nullptr;//receiver->GetSmartsuit(Controller->suitname);
 	if (!data) 
 	{
 		//UE_LOG(LogTemp, Warning, TEXT("No data for %s"), *Controller->suitname);
@@ -1132,95 +1133,7 @@ FLiveLinkSubjectName FSmartsuitPoseNode::GetLiveLinkSubjectName()
 #endif
 }
 
-//USmartsuitBodyMapData::USmartsuitBodyMapData()
-//{
-//	hip = "hip";
-//	stomach = "stomach";
-//	chest = "chest";
-//	neck = "neck";
-//	head = "head";
-//	leftShoulder = "leftShoulder";
-//	leftArm = "leftArm";
-//	leftForearm = "leftForearm";
-//	leftHand = "leftHand";
-//	rightShoulder = "rightShoulder";
-//	rightArm = "rightArm";
-//	rightForearm = "rightForearm";
-//	rightHand = "rightHand";
-//	leftUpleg = "leftUpleg";
-//	leftLeg = "leftLeg";
-//	leftFoot = "leftFoot";
-//	leftToe = "leftToe";
-//	rightUpleg = "rightUpleg";
-//	rightLeg = "rightLeg";
-//	rightFoot = "rightFoot";
-//	rightToe = "rightToe";
-//	leftThumbProximal = "leftThumbProximal";
-//	leftThumbMedial = "leftThumbMedial";
-//	leftThumbDistal = "leftThumbDistal";
-//	leftThumbTip = "leftThumbTip";
-//	leftIndexProximal = "leftIndexProximal";
-//	leftIndexMedial = "leftIndexMedial";
-//	leftIndexDistal = "leftIndexDistal";
-//	leftIndexTip = "leftIndexTip";
-//	leftMiddleProximal = "leftMiddleProximal";
-//	leftMiddleMedial = "leftMiddleMedial";
-//	leftMiddleDistal = "leftMiddleDistal";
-//	leftMiddleTip = "leftMiddleTip";
-//	leftRingProximal = "leftRingProximal";
-//	leftRingMedial = "leftRingMedial";
-//	leftRingDistal = "leftRingDistal";
-//	leftRingTip = "leftRingTip";
-//	leftLittleProximal = "leftLittleProximal";
-//	leftLittleMedial = "leftLittleMedial";
-//	leftLittleDistal = "leftLittleDistal";
-//	leftLittleTip = "leftLittleTip";
-//	rightThumbProximal = "rightThumbProximal";
-//	rightThumbMedial = "rightThumbMedial";
-//	rightThumbDistal = "rightThumbDistal";
-//	rightThumbTip = "rightThumbTip";
-//	rightIndexProximal = "rightIndexProximal";
-//	rightIndexMedial = "rightIndexMedial";
-//	rightIndexDistal = "rightIndexDistal";
-//	rightIndexTip = "rightIndexTip";
-//	rightMiddleProximal = "rightMiddleProximal";
-//	rightMiddleMedial = "rightMiddleMedial";
-//	rightMiddleDistal = "rightMiddleDistal";
-//	rightMiddleTip = "rightMiddleTip";
-//	rightRingProximal = "rightRingProximal";
-//	rightRingMedial = "rightRingMedial";
-//	rightRingDistal = "rightRingDistal";
-//	rightRingTip = "rightRingTip";
-//	rightLittleProximal = "rightLittleProximal";
-//	rightLittleMedial = "rightLittleMedial";
-//	rightLittleDistal = "rightLittleDistal";
-//	rightLittleTip = "rightLittleTip";
-//}
 
-//FName URokokoBodyMapData::GetRemappedBoneName_Implementation(FName CurveName) const
-//{
-//	if (auto RemappedName = NameMapping.Find(CurveName))
-//	{
-//		return *RemappedName;
-//	}
-//	return "";
-//}
-//
-//void URokokoBodyMapData::Initialize()
-//{
-//	InitializeTMap();
-//}
-PRAGMA_DISABLE_OPTIMIZATION
-FName URokokoBodyMapData::GetRemappedBoneName_Implementation(FName CurveName) const
-{
-	if (auto RemappedName = NameMapping.Find(CurveName))
-	{
-		return *RemappedName;
-	}
-	
-	return "";
-}
-PRAGMA_ENABLE_OPTIMIZATION
 void FSmartsuitPoseNode::OnInitializeAnimInstance(const FAnimInstanceProxy* InProxy, const UAnimInstance* InAnimInstance)
 {
 	//CurrentRetargetAsset = nullptr;
