+++ Plugins/Smartsuit/Source/Smartsuit/Public/SmartsuitPoseNode.h
@@ -2,11 +2,11 @@
 
 #pragma once
 
-//#include "UObject/NoExportTypes.h"
-//#include "Engine.h"
-//#include "AnimNode_SkeletalControlBase.h"
 #include "Runtime/AnimGraphRuntime/Public/BoneControllers/AnimNode_SkeletalControlBase.h"
 #include "LiveLinkClientReference.h"
+#include "LiveLinkCustomVersion.h"
+#include "LiveLinkRemapAsset.h"
+#include "LiveLinkTypes.h"
 #include "SmartsuitReceiver.h"
 #include "SmartsuitController.h"
 #include "SmartsuitTPose.h"
@@ -14,472 +14,6 @@
 
 class ILiveLinkClient;
 
-UCLASS(BlueprintType)
-class SMARTSUIT_API URokokoBodyMapData : public ULiveLinkRemapAsset
-{
-	GENERATED_BODY()
-
-	URokokoBodyMapData()
-	{
-		hip = "hip";
-
-		stomach = "stomach";
-
-		chest = "chest";
-
-		neck = "neck";
-
-		head = "head";
-
-		leftShoulder = "leftShoulder";
-
-		leftArm = "leftArm";
-
-		leftForearm = "leftForearm";
-
-		leftHand = "leftHand";
-
-		rightShoulder = "rightShoulder";
-
-		rightArm = "rightArm";
-
-		rightForearm = "rightForearm";
-
-		rightHand = "rightHand";
-
-		leftUpleg = "leftUpleg";
-
-		leftLeg = "leftLeg";
-
-		leftFoot = "leftFoot";
-
-		leftToe = "leftToe";
-
-		rightUpleg = "rightUpleg";
-
-		rightLeg = "rightLeg";
-
-		rightFoot = "rightFoot";
-
-		rightToe = "rightToe";
-
-		leftThumbProximal = "leftThumbProximal";
-
-		leftThumbMedial = "leftThumbMedial";
-
-		leftThumbDistal = "leftThumbDistal";
-
-		leftThumbTip = "leftThumbTip";
-
-		leftIndexProximal = "leftIndexProximal";
-
-		leftIndexMedial = "leftIndexMedial";
-
-		leftIndexDistal = "leftIndexDistal";
-
-		leftIndexTip = "leftIndexTip";
-
-		leftMiddleProximal = "leftMiddleProximal";
-
-		leftMiddleMedial = "leftMiddleMedial";
-
-		leftMiddleDistal = "leftMiddleDistal";
-
-		leftMiddleTip = "leftMiddleTip";
-
-		leftRingProximal = "leftRingProximal";
-
-		leftRingMedial = "leftRingMedial";
-
-		leftRingDistal = "leftRingDistal";
-
-		leftRingTip = "leftRingTip";
-
-		leftLittleProximal = "leftLittleProximal";
-
-		leftLittleMedial = "leftLittleMedial";
-
-		leftLittleDistal = "leftLittleDistal";
-
-		leftLittleTip = "leftLittleTip";
-
-		rightThumbProximal = "rightThumbProximal";
-
-		rightThumbMedial = "rightThumbMedial";
-
-		rightThumbDistal = "rightThumbDistal";
-
-		rightThumbTip = "rightThumbTip";
-
-		rightIndexProximal = "rightIndexProximal";
-
-		rightIndexMedial = "rightIndexMedial";
-
-		rightIndexDistal = "rightIndexDistal";
-
-		rightIndexTip = "rightIndexTip";
-
-		rightMiddleProximal = "rightMiddleProximal";
-
-		rightMiddleMedial = "rightMiddleMedial";
-
-		rightMiddleDistal = "rightMiddleDistal";
-
-		rightMiddleTip = "rightMiddleTip";
-
-		rightRingProximal = "rightRingProximal";
-
-		rightRingMedial = "rightRingMedial";
-
-		rightRingDistal = "rightRingDistal";
-
-		rightRingTip = "rightRingTip";
-
-		rightLittleProximal = "rightLittleProximal";
-
-		rightLittleMedial = "rightLittleMedial";
-
-		rightLittleDistal = "rightLittleDistal";
-
-		rightLittleTip = "rightLittleTip";
-
-
-	}
-public:
-
-	FName GetRemappedBoneName_Implementation(FName CurveName) const override;
-	// {
-	// 	if (auto RemappedName = NameMapping.Find(CurveName))
-	// 	{
-	// 		return *RemappedName;
-	// 	}
-	// 	return "";
-	// }
-
-	virtual void Initialize() override
-	{
-		InitializeTMap();
-	}
-
-	UFUNCTION(BlueprintCallable, Category = BoneRemapping)
-	void InitializeTMap()
-	{
-		NameMapping.Empty();
-		
-		NameMapping.Add("hip", hip);
-
-		NameMapping.Add("stomach", stomach);
-
-		NameMapping.Add("chest", chest);
-
-		NameMapping.Add("neck", neck);
-
-		NameMapping.Add("head", head);
-
-		NameMapping.Add("leftShoulder", leftShoulder);
-
-		NameMapping.Add("leftArm", leftArm);
-
-		NameMapping.Add("leftForearm", leftForearm);
-
-		NameMapping.Add("leftHand", leftHand);
-
-		NameMapping.Add("rightShoulder", rightShoulder);
-
-		NameMapping.Add("rightArm", rightArm);
-
-		NameMapping.Add("rightForearm", rightForearm);
-
-		NameMapping.Add("rightHand", rightHand);
-
-		NameMapping.Add("leftUpleg", leftUpleg);
-
-		NameMapping.Add("leftLeg", leftLeg);
-
-		NameMapping.Add("leftFoot", leftFoot);
-
-		NameMapping.Add("leftToe", leftToe);
-
-		NameMapping.Add("rightUpleg", rightUpleg);
-
-		NameMapping.Add("rightLeg", rightLeg);
-
-		NameMapping.Add("rightFoot", rightFoot);
-
-		NameMapping.Add("rightToe", rightToe);
-
-		NameMapping.Add("leftThumbProximal", leftThumbProximal);
-
-		NameMapping.Add("leftThumbMedial", leftThumbMedial);
-
-		NameMapping.Add("leftThumbDistal", leftThumbDistal);
-
-		NameMapping.Add("leftThumbTip", leftThumbTip);
-
-		NameMapping.Add("leftIndexProximal", leftIndexProximal);
-
-		NameMapping.Add("leftIndexMedial", leftIndexMedial);
-
-		NameMapping.Add("leftIndexDistal", leftIndexDistal);
-
-		NameMapping.Add("leftIndexTip", leftIndexTip);
-
-		NameMapping.Add("leftMiddleProximal", leftMiddleProximal);
-
-		NameMapping.Add("leftMiddleMedial", leftMiddleMedial);
-
-		NameMapping.Add("leftMiddleDistal", leftMiddleDistal);
-
-		NameMapping.Add("leftMiddleTip", leftMiddleTip);
-
-		NameMapping.Add("leftRingProximal", leftRingProximal);
-
-		NameMapping.Add("leftRingMedial", leftRingMedial);
-
-		NameMapping.Add("leftRingDistal", leftRingDistal);
-
-		NameMapping.Add("leftRingTip", leftRingTip);
-
-		NameMapping.Add("leftLittleProximal", leftLittleProximal);
-
-		NameMapping.Add("leftLittleMedial", leftLittleMedial);
-
-		NameMapping.Add("leftLittleDistal", leftLittleDistal);
-
-		NameMapping.Add("leftLittleTip", leftLittleTip);
-
-		NameMapping.Add("rightThumbProximal", rightThumbProximal);
-
-		NameMapping.Add("rightThumbMedial", rightThumbMedial);
-
-		NameMapping.Add("rightThumbDistal", rightThumbDistal);
-
-		NameMapping.Add("rightThumbTip", rightThumbTip);
-
-		NameMapping.Add("rightIndexProximal", rightIndexProximal);
-
-		NameMapping.Add("rightIndexMedial", rightIndexMedial);
-
-		NameMapping.Add("rightIndexDistal", rightIndexDistal);
-
-		NameMapping.Add("rightIndexTip", rightIndexTip);
-
-		NameMapping.Add("rightMiddleProximal", rightMiddleProximal);
-
-		NameMapping.Add("rightMiddleMedial", rightMiddleMedial);
-
-		NameMapping.Add("rightMiddleDistal", rightMiddleDistal);
-
-		NameMapping.Add("rightMiddleTip", rightMiddleTip);
-
-		NameMapping.Add("rightRingProximal", rightRingProximal);
-
-		NameMapping.Add("rightRingMedial", rightRingMedial);
-
-		NameMapping.Add("rightRingDistal", rightRingDistal);
-
-		NameMapping.Add("rightRingTip", rightRingTip);
-
-		NameMapping.Add("rightLittleProximal", rightLittleProximal);
-
-		NameMapping.Add("rightLittleMedial", rightLittleMedial);
-
-		NameMapping.Add("rightLittleDistal", rightLittleDistal);
-
-		NameMapping.Add("rightLittleTip", rightLittleTip);
-
-
-
-	}
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		TMap<FName, FName> NameMapping;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName hip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName stomach;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName chest;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName neck;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName head;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftShoulder;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftArm;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftForearm;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftHand;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightShoulder;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightArm;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightForearm;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightHand;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftUpleg;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftLeg;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftFoot;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftToe;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightUpleg;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightLeg;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightFoot;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightToe;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftThumbProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftThumbMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftThumbDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftThumbTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftIndexProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftIndexMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftIndexDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftIndexTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftMiddleProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftMiddleMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftMiddleDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftMiddleTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftRingProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftRingMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftRingDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftRingTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftLittleProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftLittleMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftLittleDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName leftLittleTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightThumbProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightThumbMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightThumbDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightThumbTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightIndexProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightIndexMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightIndexDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightIndexTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightMiddleProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightMiddleMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightMiddleDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightMiddleTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightRingProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightRingMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightRingDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightRingTip;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightLittleProximal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightLittleMedial;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightLittleDistal;
-
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = BoneRemapping)
-		FName rightLittleTip;
-
-
-};
 
 /*! \brief Bone reference mapping that is used to map bones between Smartsuit sensors and skeleton.*/
 USTRUCT()
@@ -674,198 +208,6 @@ struct SMARTSUIT_API FSmartsuitBodyMap
 	FBoneReference rightLittleTip;
 };
 
-//UCLASS(BlueprintType)
-//class SMARTSUIT_API USmartsuitBodyMapData : public UDataAsset
-//{
-//	GENERATED_BODY()
-//
-//	USmartsuitBodyMapData();
-//public:
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName hip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName stomach;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName chest;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName neck;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName head;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftShoulder;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftArm;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftForearm;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftHand;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightShoulder;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightArm;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightForearm;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightHand;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftUpleg;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftLeg;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftFoot;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftToe;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightUpleg;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightLeg;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightFoot;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightToe;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftThumbProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftThumbMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftThumbDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftThumbTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftIndexProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftIndexMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftIndexDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftIndexTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftMiddleProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftMiddleMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftMiddleDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftMiddleTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftRingProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftRingMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftRingDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftRingTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftLittleProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftLittleMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftLittleDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName leftLittleTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightThumbProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightThumbMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightThumbDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightThumbTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightIndexProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightIndexMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightIndexDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightIndexTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightMiddleProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightMiddleMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightMiddleDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightMiddleTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightRingProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightRingMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightRingDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightRingTip;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightLittleProximal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightLittleMedial;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightLittleDistal;
-//
-//	UPROPERTY(EditAnywhere, Category = SkeletalControl)
-//	FName rightLittleTip;
-//};
-
 
 /*! \brief An animation controller which poses a humanoid character to the pose as received from the Smartsuit.
 *
@@ -881,9 +223,6 @@ struct SMARTSUIT_API FSmartsuitPoseNode : public FAnimNode_SkeletalControlBase
 	UPROPERTY(/*EditAnywhere, BlueprintReadWrite, Category = SmartsuitAnimationSetup, meta = (NeverAsPin, ToolTip = "A mapping between the Smartsuit expected bone names and the bones that will animate from the Smartsuit component. Every bone is required for the animation to work properly.")*/)
 	FSmartsuitBodyMap BoneMap;
 
-	//UPROPERTY(EditAnywhere, Category = SmartsuitAnimationSetup)
-	//USmartsuitBodyMapData* Bone_Map_Override_OLD;
-
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, NoClear, Category = Retarget, meta = (PinShownByDefault))
 	TSubclassOf<ULiveLinkRemapAsset> RetargetAsset;
 
@@ -892,23 +231,8 @@ struct SMARTSUIT_API FSmartsuitPoseNode : public FAnimNode_SkeletalControlBase
 
 	void OnInitializeAnimInstance(const FAnimInstanceProxy* InProxy, const UAnimInstance* InAnimInstance) override;
 
-	/** The smartsuit controller to use. This defines which Smartsuit to use for the animation. */
-	//UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SmartsuitAnimationSetup, meta = (AlwaysAsPin, ToolTip = "The SmartsuitController that will be used to animate this character. This is required for the animation to work."))
-	//ASmartsuitController *Controller;
-
-	//UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SourceData, meta = (PinShownByDefault))
-	//FLiveLinkSubjectName LiveLinkSubjectName;
-
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SourceData, meta = (PinShownByDefault))
 	FName RokokoActorName;
-
-	///** Indicates weither the character should maintain it's XY position or override it from the position of the Smartsuit. If this is true then the character will have an offset from the actual Smartsuit position so he remains in the same position when the game begins. */
-	//UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SmartsuitAnimationSetup, meta = (PinHiddenByDefault, ToolTip = "When this is True the animation will keep the starting position in the XY-plane of the character when the game starts. This way the character will not jump out of his initial position. Its Z position however will be identical to the one received from the Smartsuit."))
-	//bool RelativeToStart;
-
-	/** Indicates if the character should scale to match the Smartsuit dimensions. This makes the character motion to be 1-1 with the player, but it may cause skinning issues with the character. */
-	//UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = SmartsuitAnimationSetup, meta = (NeverAsPin, ToolTip = "Scale bones to match the BodyModel dimensions. This will make the animation look more natural on the character but may cause unusual effects on the mesh."))
-	//	bool ScaleBones;
 
 	/// @private
 	SmartsuitTPose TPose;
